+++ openMSXbatoken.py
@@ -1,6 +1,7 @@
+#!/usr/bin/env python3
 """
 openMSX Basic (de)Tokenizer
-v1.2
+v1.3
 Uses openMSX to tokenize ASCII MSX Basic or vice-versa.
 
 Copyright (C) 2019-2020 - Fred Rique (farique)
@@ -18,20 +19,15 @@ Convert modern MSX Basic Dignified to traditional MSX Basic format.
 openmsxbatoken.py <source> <destination> [args...]
 openmsxbatoken.py -h for help.
 
-New: 26-1-2020 v1.1
-    Significant code rewrite to bring it to the Badig standard.
-    Fully integrated with the Badig ecosystem.
-        Can be automatically called by the build system on MSX Sublime Tools and from MSX Basic Dignified.
-    Created an .ini file with file_load, file_save, machine_name, disk_ext_name, output_format, delete_original, verbose_level, openmsx_filepath.
-    Verbose level and log output created to follow the Badig standard.
-    Added -do (delete original) argument to delete the original file after the conversion.
-    Added -of (output file) to indicate the format to save: tokenized or ASCII.
-    Added -fb (from build)
-    Removed -asc (replaced by -of)
-    Warning if a path is issued to the destination file. The path will be removed, the destination is always saved on the same folder (the mounted MSX disk) as the source.
-    Throw an error if destination is the same as the source
-    Will replace spaces on file names with an _ to conform to the MSX disk specification.
-    Better error handling.
+New: 1.3 14-02-2020
+    Python 3.8.
+    Better subprocess call and IO handling.
+    Improved verbose output.
+    Changed -fb to -frb.
+    Warning issued if didn' delete original.
+    Fixed bug and better handling when trying to load or save files with spaces and more than 8 characters.
+        Files opened on openMSX now are internally cropped to 8 char and have spaces replaced with _
+        Error if conflicting file names due to disk format limitations.
 """
 
 # for i in *.bas; do python openmsxbatoken.py "$i" -asc; done
@@ -40,7 +36,7 @@ New: 26-1-2020 v1.1
 import os.path
 import argparse
 import subprocess
-import ConfigParser
+import configparser
 from os import remove as osremove
 
 file_load = ''              # Source file
@@ -70,23 +66,23 @@ def show_log(line, text, level, **kwargs):
     line_number = line
 
     if verbose_level >= level:
-        print bullets[bullet] + display_file_name + line_number + text
+        print(bullets[bullet] + display_file_name + line_number + text)
         # print bullets[bullet] + text
 
     if bullet == 1:
         if proc:
             proc.stdin.write('<command>type_via_keybuf poke-2,0\\r</command>')
         if is_from_build:
-            print '    Tokenizing_aborted'
+            print('    Tokenizing_aborted')
         else:
-            print '    Execution_stoped'
-            print
+            print('    Execution_stoped')
+            print()
         raise SystemExit(0)
 
 
 local_path = os.path.split(os.path.abspath(__file__))[0] + '/'
 if os.path.isfile(local_path + 'openMSXBatoken.ini'):
-    config = ConfigParser.ConfigParser()
+    config = configparser.ConfigParser()
     config.sections()
     try:
         config.read(local_path + 'openMSXBatoken.ini')
@@ -98,7 +94,7 @@ if os.path.isfile(local_path + 'openMSXBatoken.ini'):
         delete_original = config.getboolean('DEFAULT', 'delete_original') if config.get('DEFAULT', 'delete_original') else delete_original
         verbose_level = config.getint('DEFAULT', 'verbose_level') if config.get('DEFAULT', 'verbose_level') else verbose_level
         openmsx_filepath = config.get('DEFAULT', 'openmsx_filepath') if config.get('DEFAULT', 'openmsx_filepath') else openmsx_filepath
-    except (ValueError, ConfigParser.NoOptionError) as e:
+    except (ValueError, configparser.NoOptionError) as e:
         show_log('', 'openMSXBatoken.ini: ' + str(e), 1)
 
 parser = argparse.ArgumentParser(description='Use openMSX to convert between ASCII and tokenized MSX Basic.')
@@ -107,7 +103,7 @@ parser.add_argument("file_save", nargs='?', default=file_save, help='The file to
 parser.add_argument("-of", default=output_format, choices=['t', 'T', 'a', 'A'], help="Tokenized or ASCII output: t-tokenized(def) a-ASCII")
 parser.add_argument("-do", default=delete_original, action='store_true', help="Delete original file after conversion")
 parser.add_argument("-vb", default=verbose_level, type=int, help="Verbosity level: 0 silent, 1 errors, 2 +warnings, 3 +steps(def), 4 +detalis")
-parser.add_argument("-fb", default=is_from_build, action='store_true', help="Tell it is running from a build system")
+parser.add_argument("-frb", default=is_from_build, action='store_true', help="Tell it is running from a build system")
 args = parser.parse_args()
 
 file_load = args.file_load
@@ -115,7 +111,7 @@ file_save = args.file_save
 output_format = args.of.upper()
 delete_original = args.do
 verbose_level = args.vb
-is_from_build = args.fb
+is_from_build = args.frb
 
 show_log('', '', 3, bullet=0)
 
@@ -132,7 +128,7 @@ if openmsx_filepath == '':
     openmsx_filepath = local_path + 'openMSX.app'
 if machine_name != '':
     using_machine = machine_name
-    machine_name = '-machine ' + machine_name
+    machine_name = ['-machine', machine_name]
 disk_ext = disk_ext_name.split(':')
 disk_ext_name = disk_ext[0].strip()
 if len(disk_ext) > 1:
@@ -147,90 +143,126 @@ if output_format == 'A':
     save_argument = '",a'
 if file_save == '':
     file_save = os.path.basename(file_load)
-    file_save = os.path.splitext(file_save)[0][0:8] + save_extension
+    file_save = os.path.splitext(file_save)[0] + save_extension
 disk_path = os.path.dirname(file_load)
 disk_path = local_path if disk_path == '' else disk_path
 file_load = os.path.basename(file_load)
 
-disk_path = disk_path.replace(' ', r'\ ')
-file_load = file_load.replace(' ', '_')
-file_save = file_save.replace(' ', '_')
+crop_load = os.path.splitext(file_load)[0][0:8] + os.path.splitext(file_load)[1]
+crop_save = os.path.splitext(file_save)[0][0:8] + os.path.splitext(file_save)[1]
 
 if save_rest:
     show_log('', ' '.join(['destination_path_removed', save_rest]), 2)
 
-if file_save == file_load:
-    show_log('', ' '.join(['destination_same_as_source', file_save]), 1)  # Exit
+if crop_save == file_load:
+    show_log('', ' '.join(['destination_same_as_source', crop_save]), 1)  # Exit
+
+list_dir = os.listdir(disk_path)
+list_load = [x for x in list_dir if
+             x.lower() != file_load.lower() and
+             os.path.splitext(x)[0][0:8].replace(' ', '_').lower() +
+             os.path.splitext(x)[1].replace(' ', '_').lower() ==
+             crop_load.lower()]
+
+list_save = [x for x in list_dir if
+             x.lower() != file_load.lower() and
+             os.path.splitext(x)[0][0:8].replace(' ', '_').lower() +
+             os.path.splitext(x)[1].replace(' ', '_').lower() ==
+             crop_load.lower() and
+             len(os.path.splitext(x)[0][0:8]) > 8]
+
+list_all = list_load.extend(list_save)
+
+if list_all:
+    show_log('', ' '.join(['MSX_disk_name_format_conflict', ', '.join(list_all)]), 1)  # Exit
+
+disk_path = disk_path.replace(' ', r'\ ')
+crop_load = crop_load.replace(' ', '_')
+crop_save = crop_save.replace(' ', '_')
 
 
-def output(show_output, step):
+def output(show_output, has_input, step):
     if show_output:
-        # proc.stdin.flush()
-        log_out = proc.stdout.readline().rstrip()
+        log_out = proc.stdout.readline().rstrip() if has_input else ''
         log_out = log_out.replace('&quot;', '"')
+        log_out = log_out.replace('&apos;', "'")
         if '"nok"' in log_out or ' error: ' in log_out:
             log_out = log_out.replace('<reply result="nok">', '')
             proc.stdin.write('<command>quit</command>')
             show_log('', ''.join([step]), 3)
+            if 'invalid command name "ext' in log_out:
+                show_log('', ''.join(['Machine probably missing a slot']), 2)
             show_log('', ''.join([log_out]), 1)  # Exit
         elif '<log level="warning">' in log_out:
             log_warning = log_out.replace('<log level="warning">', '')
             log_warning = log_warning.replace('</log>', '')
             log_out = log_out.split('<log')[0]
             log_comma = '' if log_out == '' else ': '
-            show_log('', ''.join([step, log_comma, log_out]), 3)
+            if step + log_comma + log_out != '':
+                show_log('', ''.join([step, log_comma, log_out]), 3)
             show_log('', ''.join([log_warning]), 2)
+            output(show_output, True, '')
         else:
             log_out = log_out.replace('<openmsx-output>', '')
             log_out = log_out.replace('</openmsx-output>', '')
             log_out = log_out.replace('<reply result="ok">', '')
             log_out = log_out.replace('</reply>', '')
             log_comma = '' if log_out == '' else ': '
-            show_log('', ''.join([step, log_comma, log_out]), 3)
+            if step + log_comma + log_out != '':
+                show_log('', ''.join([step, log_comma, log_out]), 3)
 
 
-cmd = (openmsx_filepath + '/contents/macos/openmsx ' + machine_name + ' -control stdio')
-proc = subprocess.Popen([cmd], shell=True, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+cmd = [openmsx_filepath + '/contents/macos/openmsx', '-control', 'stdio']
+if machine_name != '':
+    cmd.extend(machine_name)
 
-output(show_output, 'openMSX initialized as ' + using_machine)
+proc = subprocess.Popen(cmd, bufsize=1, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, encoding='utf-8')
 
-# proc.stdin.write('<command>set renderer SDL</command>')
-# output(show_output, 'Show screen')
+endline = '\r\n'
 
-proc.stdin.write('<command>set throttle off</command>')
-output(show_output, 'Set throttle off')
+output(show_output, True, 'openMSX initialized as ' + using_machine)
 
-proc.stdin.write('<command>debug set_watchpoint write_mem 0xfffe {[debug read "memory" 0xfffe] == 0} {quit}</command>')
-output(show_output, 'Set quit watchpoint')
+# proc.stdin.write('<command>set renderer SDL</command>' + endline)
+# output(show_output, True, 'Show screen')
+
+proc.stdin.write('<command>set throttle off</command>' + endline)
+output(show_output, True, 'Set throttle off')
+
+proc.stdin.write('<command>debug set_watchpoint write_mem 0xfffe {[debug read "memory" 0xfffe] == 0} {quit}</command>' + endline)
+output(show_output, True, 'Set quit watchpoint')
 
 if disk_ext_name != '':
-    proc.stdin.write('<command>' + disk_ext_slot + ' ' + disk_ext_name + '</command>')
-    output(show_output, 'Insert disk drive extension: ' + disk_ext_name + ' at ' + disk_ext_slot)
+    proc.stdin.write('<command>' + disk_ext_slot + ' ' + disk_ext_name + '</command>' + endline)
+    output(show_output, True, 'Insert disk drive extension: ' + disk_ext_name + ' at ' + disk_ext_slot)
 
-proc.stdin.write('<command>diska insert ' + disk_path + '</command>')
-output(show_output, 'insert folder as disk: ' + disk_path)
+proc.stdin.write('<command>diska insert ' + disk_path + '</command>' + endline)
+output(show_output, True, 'insert folder as disk: ' + disk_path)
 
-proc.stdin.write('<command>set power on</command>')
-output(show_output, 'Power on')
+proc.stdin.write('<command>set power on</command>' + endline)
+output(show_output, True, 'Power on')
 
-proc.stdin.write('<command>type_via_keybuf \\r\\r</command>')  # Disk ROM ask for date, enter twice to skip
-output(show_output, 'Press return twice')
+proc.stdin.write('<command>type_via_keybuf \\r\\r</command>' + endline)  # Disk ROM ask for date, enter twice to skip
+output(show_output, True, 'Press return twice')
 
-proc.stdin.write('<command>type_via_keybuf load"' + file_load + '\\r</command>')
-output(show_output, 'type load"' + file_load + '"')
+proc.stdin.write('<command>type_via_keybuf load"' + crop_load + '\\r</command>' + endline)
+output(show_output, True, 'type load"' + crop_load)
 
-proc.stdin.write('<command>type_via_keybuf save"' + file_save + save_argument + '\\r</command>')
-output(show_output, 'type save"' + file_save + save_argument + '"')
+proc.stdin.write('<command>type_via_keybuf save"' + crop_save + save_argument + '\\r</command>' + endline)
+output(show_output, True, 'type save"' + crop_save + save_argument)
 
-proc.stdin.write('<command>type_via_keybuf poke-2,0\\r</command>')
-output(show_output, 'Quit')
+proc.stdin.write('<command>type_via_keybuf poke-2,0\\r</command>' + endline)
+output(show_output, True, 'Quit')
 
 proc.wait()
 
-file_save_full = disk_path + '/' + file_save
-if os.path.isfile(file_save_full) and delete_original:
-    show_log('', 'Deleting source', 3)
-    show_log('', ' '.join(['delete_file:', file_load_full]), 4)
-    osremove(file_load_full)
+file_save_full = disk_path + '/' + crop_save
+if delete_original:
+    if os.path.isfile(file_save_full):
+        show_log('', 'Deleting source', 3)
+        show_log('', ' '.join(['delete_file:', file_load_full]), 4)
+        osremove(file_load_full)
+    else:
+        show_log('', ' '.join(['source_not_deleted', file_load_full]), 2)
+        show_log('', ' '.join(['converted_not_found', crop_save]), 2)
 
 show_log('', '', 3, bullet=0)